change file structure, add unit test for getEnabledQuestions()
