+++ src/applications/coronavirus-screener/lib/scrollTo.js
@@ -1,0 +1,13 @@
+import { scroller } from 'react-scroll';
+
+// scroller usage based on https://github.com/department-of-veterans-affairs/veteran-facing-services-tools/blob/master/packages/formation-react/src/components/CollapsiblePanel/CollapsiblePanel.jsx
+export function scrollTo(name) {
+  scroller.scrollTo(
+    name,
+    window.VetsGov.scroll || {
+      duration: 500,
+      delay: 2,
+      smooth: true,
+    },
+  );
+}
